backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
bcrypt==4.0.1
sentry-sdk[fastapi]==1.38.0
